feat: Create test controller
